+++ src/main/kotlin/Main.kt
@@ -1,7 +1,60 @@
-fun main(args: Array<String>) {
-    println("Hello World!")
+data class Word(
+    val original: String,
+    val translate: String,
+    var correctAnswersCount: Int = 0,
+)
 
-    // Try adding program arguments via Run/Debug configuration.
-    // Learn more about running applications: https://www.jetbrains.com/help/idea/running-applications.html.
-    println("Program arguments: ${args.joinToString()}")
+fun Question.asConsoleString(): String {
+    val variants = this.variants
+        .mapIndexed { index: Int, word: Word -> "${index + 1} - ${word.translate}" }
+        .joinToString(separator = "\n")
+    return this.correctAnswer.original + "\n" + variants + "\n0 - выйти в главное меню"
+}
+
+fun main() {
+
+    val trainer = try {
+        LearnWordsTrainer(3, 4)
+    } catch (e: Exception) {
+        println("Невозможно загрузить словарь")
+        return
+    }
+
+    while (true) {
+        println("Меню: 1 – Учить слова, 2 – Статистика, 0 – Выход")
+        when (readln().toIntOrNull()) {
+            1 -> {
+                while (true) {
+                    val question = trainer.getNextQuestion()
+                    if (question == null) {
+                        println("Вы выучили все слова")
+                        break
+                    } else {
+                        println(question.asConsoleString())
+
+                        val userAnswerInput = readln().toIntOrNull()
+                        if (userAnswerInput == 0) break
+
+                        if (trainer.checkAnswer(userAnswerInput?.minus(1))) {
+                            println("Верно!")
+                        } else println(
+                            "Не верно. Перевод слова ${question.correctAnswer.translate} - " +
+                                    question.correctAnswer.original
+                        )
+                    }
+                }
+            }
+
+            2 -> {
+                val statistics = trainer.getStatistics()
+                println(
+                    "Выучено ${statistics.countLearnedWords} из ${statistics.countAllWords} слов | " +
+                            "${statistics.learnedPercent}%"
+                )
+            }
+
+            0 -> break
+            else -> println("Введите номер пункта меню")
+        }
+    }
 }
